Move Tuple(N) into tuple_types.h
